+++ .eslintrc.js
@@ -5,7 +5,7 @@ module.exports = {
     globals: {
         'epicenter': true
     },
-    ignorePatterns: ['**/dist/**', '*.cjs'],
+    ignorePatterns: ['**/dist/**'],
     parser: '@typescript-eslint/parser',
     plugins: [
       '@typescript-eslint',