+++ src/adapters/leaderboard.ts
@@ -1,7 +1,8 @@
-import type { RoutingOptions, GenericSearchOptions } from '../utils/router';
-import type { GenericScope } from '../utils/constants';
+import type { UserSession } from 'utils/identification';
+import type { RoutingOptions, GenericSearchOptions } from 'utils/router';
+import type { GenericScope } from 'utils/constants';
 
-import { identification, Router } from '../utils';
+import { identification, Router } from 'utils';
 
 /**
  * Leaderboard API adapter
@@ -22,6 +23,7 @@ interface Leaderboard {
     lastUpdated: Date,
 }
 
+
 /**
  * Updates leaderboard information
  *
@@ -36,25 +38,22 @@ interface Leaderboard {
 export async function update(
     collection: string,
     scores: Score[],
-    scope: GenericScope,
+    scope: { userKey?: string } & GenericScope,
     optionals: {
         tags?: Tag[],
         userKey?: string,
     } & RoutingOptions = {}
 ): Promise<Leaderboard> {
-    const {
-        tags, userKey,
-        ...routingOptions
-    } = optionals;
-    const { scopeBoundary, scopeKey } = scope;
-
+    const { tags, ...routingOptions } = optionals;
+    const { scopeBoundary, scopeKey, userKey } = scope;
+    const session = identification.session as UserSession;
     return await new Router()
         .post('/leaderboard', {
             body: {
                 scope: {
                     scopeBoundary,
                     scopeKey,
-                    userKey: userKey ?? identification.session?.userKey,
+                    userKey: userKey ?? session?.userKey,
                 },
                 collection,
                 scores, tags,
@@ -64,7 +63,7 @@ export async function update(
 }
 
 /**
- * Gathers leaderboard information
+ * Gathers leaderboard information; not pageable
  *
  * Base URL: GET `https://forio.com/api/v3/{ACCOUNT}/{PROJECT}/leaderboard/{SCOPE_BOUNDARY}/{SCOPE_KEY}/{COLLECTION}`
  *
@@ -72,7 +71,7 @@ export async function update(
  * @example
  *
  * import { leaderboardAdapter } from 'epicenter';
- * const leaderboard = await leaderboardAdapter.get('myLeaderboard', {
+ * const leaderboard = await leaderboardAdapter.list('myLeaderboard', scope, {
  *      filter: [
  *          'tag.role=doctor',  // look for leaderboard entries tagged with role=doctor
  *          'score.total>0'     // where the users scored a total higher than 0
@@ -82,12 +81,33 @@ export async function update(
  *      max: 20                 // retrieve only the first 20 entries
  * });
  */
+export async function list(
+    collection: string,
+    scope: GenericScope,
+    searchOptions: GenericSearchOptions,
+    optionals: RoutingOptions = {}
+): Promise<Leaderboard[]> {
+    const { scopeBoundary, scopeKey } = scope;
+    const { filter = [], sort = [], first, max } = searchOptions;
+    const searchParams = {
+        filter: filter.join(';') || undefined,
+        sort: sort.join(';') || undefined,
+        first, max,
+    };
+
+    return await new Router()
+        .withSearchParams(searchParams)
+        .get(`/leaderboard/${scopeBoundary}/${scopeKey}/${collection}`, optionals)
+        .then(({ body }) => body);
+}
+
 export async function get(
     collection: string,
     scope: GenericScope,
     searchOptions: GenericSearchOptions,
     optionals: RoutingOptions = {}
 ): Promise<Leaderboard[]> {
+    console.warn('DEPRECATION WARNING: leaderboardAdapter.get is deprecated and will be removed with the next release. Use leaderboardAdapter.list instead.');
     const { scopeBoundary, scopeKey } = scope;
     const { filter = [], sort = [], first, max } = searchOptions;
     const searchParams = {